fix(F153): gate shadow misses behind relaxed action heuristic
P2-3 from re-review: `!hasActionKeyword` was directly equated with
shadow miss, causing pure narrative mentions ("之前 @codex 提出的方案不错")
to pollute the shadow_miss metric. Now requires RELAXED_*_ACTION_RE
match — a superset of strict handoff patterns that still filters
narrative mentions while catching broader vocab gap candidates.

Also extracts shadow detection into a2a-shadow-detection.ts to keep
a2a-mentions.ts under the 350-line file size limit (245 + 139).

Adds 2 tests: narrative mention negative case + relaxed action positive.

[宪宪/Opus-46🐾]
