+++ packages/api/src/domains/cats/services/agents/routing/a2a-mentions.ts
@@ -11,7 +11,6 @@
  * 6. 只在猫回复完整结束后解析 (由调用方保证)
  */
 
-import { createHash } from 'node:crypto';
 import type { CatId } from '@cat-cafe/shared';
 import { CAT_CONFIGS, catRegistry } from '@cat-cafe/shared';
 import { isCatAvailable } from '../../../../../config/cat-config-loader.js';
@@ -23,10 +22,10 @@ export function getMaxA2ADepth(): number {
 
 /** Max number of distinct cats a single message can @mention (F27 safety limit) */
 const MAX_A2A_MENTION_TARGETS = 2;
-const TOKEN_BOUNDARY_RE = /[\s,.:;!?()[\]{}<>，。！？、：；（）【】《》「」『』〈〉]/;
-// If the next char looks like part of a handle token, treat it as NOT a boundary.
-// This avoids prefix-matching `@opus-45` as `@opus`, while still allowing `@opus请看`.
-const HANDLE_CONTINUATION_RE = /[a-z0-9_.-]/;
+/** @internal Exported for a2a-shadow-detection.ts. */
+export const TOKEN_BOUNDARY_RE = /[\s,.:;!?()[\]{}<>，。！？、：；（）【】《》「」『』〈〉]/;
+/** @internal Exported for a2a-shadow-detection.ts. */
+export const HANDLE_CONTINUATION_RE = /[a-z0-9_.-]/;
 const LEADING_MARKDOWN_MENTION_PREFIX_RE = /^(?:(?:>\s*)|(?:[-*+]\s+)|(?:\d+[.)]\s+))+/;
 
 interface MentionPatternEntry {
@@ -160,13 +159,15 @@ export function analyzeA2AMentions(
  * Action patterns that appear immediately BEFORE @mention (e.g. "Ready for @xxx").
  * Chinese 请 uses negative lookbehind to exclude compounds (邀请 = invite, 申请 = apply).
  */
-const BEFORE_HANDOFF_RE = /(?:ready\s+for|交接给?|转给|(?<![邀申敬])请|帮)\s*$/i;
+/** @internal Exported for a2a-shadow-detection.ts. */
+export const BEFORE_HANDOFF_RE = /(?:ready\s+for|交接给?|转给|(?<![邀申敬])请|帮)\s*$/i;
 /**
  * Action patterns immediately AFTER @mention (e.g. "@xxx review").
  * English verbs use (?![a-z]) to reject continuations ("reviewed", "checklist").
  * Chinese verbs use negative lookahead to exclude completion suffixes (过/了/完/好/掉).
  */
-const AFTER_HANDOFF_RE =
+/** @internal Exported for a2a-shadow-detection.ts. */
+export const AFTER_HANDOFF_RE =
   /^\s*(?:(?:review|check|fix|merge)(?![a-z])|(?:确认|处理|来处理|来看)(?![过了完好掉])|看一?下|帮忙|请(?![教示假求问]))/i;
 
 export function detectInlineActionMentions(
@@ -239,111 +240,6 @@ export function detectInlineActionMentions(
   return found;
 }
 
-// --- F479: Shadow detection (in-process, metadata-only) ---
-
-/** Shadow miss metadata — no raw text, per data minimization (mindfn #479). */
-export interface ShadowMiss {
-  readonly catId: CatId;
-  /** SHA-256 of the line context, truncated to 16 hex chars. */
-  readonly contextHash: string;
-  /** Length of the raw line. */
-  readonly contextLength: number;
-}
-
-export interface ShadowDetectionResult {
-  readonly strictHits: InlineActionMention[];
-  readonly shadowMisses: ShadowMiss[];
-  /** Count of mentions skipped because routedSet already covered them. */
-  readonly routedSetSkips: number;
-}
-
-function hashContext(line: string): string {
-  return createHash('sha256').update(line).digest('hex').slice(0, 16);
-}
-
-/**
- * F479: Run strict detection + relaxed shadow detection in one pass.
- *
- * Shadow detection = "any inline @mention not caught by strict detection and
- * not in routedSet". These are vocab gap candidates where the action keyword
- * is present but not in our regex.
- */
-export function detectInlineActionMentionsWithShadow(
-  text: string,
-  currentCatId?: CatId,
-  routedMentions?: CatId[],
-): ShadowDetectionResult {
-  if (!text) return { strictHits: [], shadowMisses: [], routedSetSkips: 0 };
-
-  const strictHits = detectInlineActionMentions(text, currentCatId, routedMentions);
-
-  // Relaxed pass: find ALL inline @mentions, classify per occurrence.
-  // P2-1 fix: check action keywords per occurrence instead of global catId diff.
-  // P2-2 fix: only count routedSetSkip when action keyword present, dedup per (catId, line).
-  const stripped = text.replace(/```[\s\S]*?```/g, '');
-  const allConfigs = Object.keys(catRegistry.getAllConfigs()).length > 0 ? catRegistry.getAllConfigs() : CAT_CONFIGS;
-
-  const entries: MentionPatternEntry[] = [];
-  for (const [id, config] of Object.entries(allConfigs)) {
-    if (currentCatId && id === currentCatId) continue;
-    if (!isCatAvailable(id)) continue;
-    for (const pattern of config.mentionPatterns) {
-      entries.push({ catId: id as CatId, pattern: pattern.toLowerCase() });
-    }
-  }
-  entries.sort((a, b) => b.pattern.length - a.pattern.length);
-
-  const routedSet = new Set(routedMentions ?? []);
-  const shadowMisses: ShadowMiss[] = [];
-  let routedSetSkips = 0;
-
-  for (const rawLine of stripped.split(/\r?\n/)) {
-    const trimmed = rawLine.trimStart();
-    const normalized = trimmed.toLowerCase();
-    if (normalized.startsWith('>')) continue;
-
-    // Per-line dedup: same cat on same line only counted once per category
-    const lineShadowSeen = new Set<string>();
-    const lineRoutedSkipSeen = new Set<string>();
-
-    for (const entry of entries) {
-      let searchFrom = 0;
-      while (searchFrom < normalized.length) {
-        const idx = normalized.indexOf(entry.pattern, searchFrom);
-        if (idx < 0) break;
-        searchFrom = idx + 1;
-        if (idx === 0) continue; // line-start → handled by parseA2AMentions
-        if (HANDLE_CONTINUATION_RE.test(normalized[idx - 1]!)) continue;
-        const charAfter = normalized[idx + entry.pattern.length];
-        const isBoundary = !charAfter || TOKEN_BOUNDARY_RE.test(charAfter) || !HANDLE_CONTINUATION_RE.test(charAfter);
-        if (!isBoundary) continue;
-
-        // Per-occurrence action keyword check
-        const before = normalized.slice(0, idx);
-        const after = normalized.slice(idx + entry.pattern.length);
-        const hasActionKeyword = BEFORE_HANDOFF_RE.test(before) || AFTER_HANDOFF_RE.test(after);
-
-        if (routedSet.has(entry.catId)) {
-          // P2-2: only count skip when action keyword present (genuine overlap)
-          if (hasActionKeyword && !lineRoutedSkipSeen.has(entry.catId)) {
-            lineRoutedSkipSeen.add(entry.catId);
-            routedSetSkips++;
-          }
-          break;
-        }
-        // Shadow miss = inline @mention WITHOUT action keyword (vocab gap candidate)
-        if (!hasActionKeyword && !lineShadowSeen.has(entry.catId)) {
-          lineShadowSeen.add(entry.catId);
-          shadowMisses.push({
-            catId: entry.catId,
-            contextHash: hashContext(rawLine.trim()),
-            contextLength: rawLine.trim().length,
-          });
-        }
-        break;
-      }
-    }
-  }
-
-  return { strictHits, shadowMisses, routedSetSkips };
-}
+// --- F479: Shadow detection — extracted to a2a-shadow-detection.ts ---
+export type { ShadowDetectionResult, ShadowMiss } from './a2a-shadow-detection.js';
+export { detectInlineActionMentionsWithShadow } from './a2a-shadow-detection.js';
